Remove disqus and make home page look slightly better
