+++ gatsby/src/components/Home/index.tsx
@@ -18,6 +18,14 @@ interface SectionProps {
 const Section = styled.section<SectionProps>`
   width: 50%;
 
+  & .listening-img {
+    width: 100%;
+  }
+
+  @media (max-width: 768px) {
+    width: 100%;
+  }
+
   ${props =>
     props.fullWidth &&
     css`
@@ -28,7 +36,22 @@ const Section = styled.section<SectionProps>`
 const Home: React.FC = () => (
   <Wrapper className="about">
     <Section className="feeling">
-      <Daylio />
+      <h2>Feeling</h2>
+      <Daylio variant="home" />
+    </Section>
+    <Section className="listening">
+      <h2>Listening</h2>
+      <a href="https://www.last.fm/user/eli_pwnd">
+        <img
+          className="listening-img"
+          src="/img/last.fm.jpg"
+          alt="My top 9 albums for the past 7 days"
+        />
+      </a>
+    </Section>
+    <Section className="coding">
+      <h2>Coding</h2>
+      <GitHubCalendar username="eligundry" dateFormat="YYYY-MM-DD" />
     </Section>
     <Section className="reading">
       <h2>Reading</h2>
@@ -37,16 +60,6 @@ const Home: React.FC = () => (
         apiKey="TSX2BO7dC8AMZstT2H6MBg"
         shelf="currently-reading"
       />
-    </Section>
-    <Section className="listening">
-      <h2>Listening</h2>
-      <a href="https://www.last.fm/user/eli_pwnd">
-        <img src="/img/last.fm.jpg" alt="My top 9 albums for the past 7 days" />
-      </a>
-    </Section>
-    <Section className="coding">
-      <h2>Coding</h2>
-      <GitHubCalendar username="eligundry" dateFormat="YYYY-MM-DD" />
     </Section>
     <Section className="tweets" fullWidth>
       <h2>Tweeting</h2>
