--- gatsby/src/templates/listing.css
@@ -1,16 +1,0 @@
-.listing-container {
-  min-height: 100%;
-  min-width: 60%;
-  margin-left: auto;
-  margin-right: auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.paging-container {
-  min-width: 100%;
-  display: flex;
-  justify-content: space-between;
-}
