+++ gatsby/src/templates/listing.jsx
@@ -5,35 +5,8 @@ import Layout from '../layout'
 import PostListing from '../components/PostListing/PostListing'
 import SEO from '../components/SEO/SEO'
 import config from '../../data/SiteConfig'
-import './listing.css'
 
 class Listing extends React.Component {
-  renderPaging() {
-    const { currentPageNum, pageCount } = this.props.pageContext
-    const prevPage = currentPageNum - 1 === 1 ? '/' : `/${currentPageNum - 1}/`
-    const nextPage = `/${currentPageNum + 1}/`
-    const isFirstPage = currentPageNum === 1
-    const isLastPage = currentPageNum === pageCount
-
-    return (
-      <div className="paging-container">
-        {!isFirstPage && <Link to={prevPage}>Previous</Link>}
-        {[...Array(pageCount)].map((_val, index) => {
-          const pageNum = index + 1
-          return (
-            <Link
-              key={`listing-page-${pageNum}`}
-              to={pageNum === 1 ? '/' : `/${pageNum}/`}
-            >
-              {pageNum}
-            </Link>
-          )
-        })}
-        {!isLastPage && <Link to={nextPage}>Next</Link>}
-      </div>
-    )
-  }
-
   render() {
     const postEdges = this.props.data.allMarkdownRemark.edges
 
@@ -45,7 +18,6 @@ class Listing extends React.Component {
             <SEO />
             <PostListing postEdges={postEdges} />
           </div>
-          {this.renderPaging()}
         </div>
       </Layout>
     )
